+++ packages/algob/src/builtin-tasks/deploy.ts
@@ -1,9 +1,9 @@
+import { BuilderError } from "@algorand-builder/runtime";
 import fs from "fs";
 import glob from "glob";
 import path from "path";
 
 import { task } from "../internal/core/config/config-env";
-import { BuilderError } from "../internal/core/errors";
 import { ERRORS } from "../internal/core/errors-list";
 import { AlgoOperator, createAlgoOperator } from "../lib/algo-operator";
 import { cmpStr } from "../lib/comparators";
