+++ packages/algob/src/internal/cli/help-printer.ts
@@ -1,3 +1,5 @@
+import { BuilderError } from "@algorand-builder/runtime";
+
 import { cmpStr } from "../../lib/comparators";
 import {
   AlgobParamDefinitions,
@@ -6,7 +8,6 @@ import {
   ParamDefinitionsMap,
   TasksMap
 } from "../../types";
-import { BuilderError } from "../core/errors";
 import { ERRORS } from "../core/errors-list";
 import { ArgumentsParser } from "./arguments-parser";
