+++ packages/algob/src/internal/core/errors.ts
@@ -1,52 +1,14 @@
-import { applyErrorMessageTemplate } from "@algorand-builder/runtime";
+import { BuilderError } from "@algorand-builder/runtime";
 import type { RequestError } from 'algosdk';
 
 import type { AnyMap } from "../../types";
 import { getClosestCallerPackage } from "../util/caller-package";
-import { ErrorDescriptor, ERRORS, getErrorCode } from "./errors-list";
+import { ERRORS } from "./errors-list";
 
 export { ERRORS }; // re-export errors-list
 
 // For an explanation about these classes constructors go to:
 // https://github.com/Microsoft/TypeScript/wiki/Breaking-Changes#extending-built-ins-like-error-array-and-map-may-no-longer-work
-
-export class BuilderError extends Error {
-  public static isBuilderError(other: any): other is BuilderError { // eslint-disable-line
-    return (
-      other !== undefined && other !== null && other._isBuilderError === true
-    );
-  }
-
-  public readonly errorDescriptor: ErrorDescriptor;
-  public readonly number: number;
-  public readonly parent?: Error;
-
-  private readonly _isBuilderError: boolean;
-
-  constructor (
-    errorDescriptor: ErrorDescriptor,
-    messageArguments: AnyMap = {},
-    parentError?: Error
-  ) {
-    const prefix = `${getErrorCode(errorDescriptor)}: `;
-
-    const formattedMessage = applyErrorMessageTemplate(
-      errorDescriptor.message,
-      messageArguments
-    );
-
-    super(prefix + formattedMessage);
-    this.errorDescriptor = errorDescriptor;
-    this.number = errorDescriptor.number;
-
-    if (parentError instanceof Error) {
-      this.parent = parentError;
-    }
-
-    this._isBuilderError = true;
-    Object.setPrototypeOf(this, BuilderError.prototype);
-  }
-}
 
 /**
  * This class is used to throw errors from algob plugins.
