+++ packages/algob/src/internal/core/plugins.ts
@@ -1,10 +1,10 @@
+import { BuilderError } from "@algorand-builder/runtime";
 import debug from "debug";
 import * as path from "path";
 import * as semver from "semver";
 
 import { StrMap } from "../../types";
 import { BuilderContext } from "../context";
-import { BuilderError } from "./errors";
 import { ERRORS } from "./errors-list";
 import { ExecutionMode, getExecutionMode } from "./execution-mode";
