+++ packages/algob/src/internal/core/runtime-environment.ts
@@ -1,3 +1,4 @@
+import { BuilderError } from "@algorand-builder/runtime";
 import debug from "debug";
 
 import type {
@@ -13,7 +14,6 @@ import type {
   TaskDefinition,
   TasksMap
 } from "../../types";
-import { BuilderError } from "./errors";
 import { ERRORS } from "./errors-list";
 import { OverriddenTaskDefinition } from "./tasks/task-definitions";
