+++ packages/algob/src/lib/script-checkpoints.ts
@@ -1,10 +1,10 @@
 
+import { BuilderError, loadFromYamlFileSilent } from "@algorand-builder/runtime";
 import deepEqual from "deep-equal";
 import * as fs from "fs";
 import path from "path";
 import YAML from "yaml";
 
-import { BuilderError } from "../internal/core/errors";
 import { ERRORS } from "../internal/core/errors-list";
 import {
   ASAInfo,
@@ -15,7 +15,6 @@ import {
   LsigInfo,
   SSCInfo
 } from "../types";
-import { loadFromYamlFileSilent } from "./files";
 
 export const scriptsDirectory = "scripts";
 const artifactsPath = "artifacts";
