+++ packages/algob/src/lib/tx.ts
@@ -1,11 +1,10 @@
 import { encodeNote, mkTransaction } from "@algorand-builder/runtime";
-import type { ExecParams, TxParams } from "@algorand-builder/runtime/build/types";
+import type { ASADef, ExecParams, TxParams } from "@algorand-builder/runtime/build/types";
 import { SignType } from "@algorand-builder/runtime/build/types";
 import algosdk, { Algodv2, SuggestedParams, Transaction } from "algosdk";
 
 import {
   AlgobDeployer,
-  ASADef,
   ASADeploymentFlags
 } from "../types";
 import { ALGORAND_MIN_TX_FEE } from "./algo-operator";
