+++ packages/algob/test/internal/deployer.ts
@@ -1,3 +1,4 @@
+import type { ASADef } from "@algorand-builder/runtime/build/types";
 import type { LogicSig } from "algosdk";
 import { assert } from "chai";
 
@@ -6,7 +7,7 @@ import { DeployerDeployMode } from "../../src/internal/deployer";
 import { DeployerConfig } from "../../src/internal/deployer_cfg";
 import { getDummyLsig } from "../../src/lib/lsig";
 import { CheckpointRepoImpl } from "../../src/lib/script-checkpoints";
-import { ASADef, ASAInfo, Checkpoints, LsigInfo, SSCInfo } from "../../src/types";
+import { ASAInfo, Checkpoints, LsigInfo, SSCInfo } from "../../src/types";
 import { expectBuilderError, expectBuilderErrorAsync } from "../helpers/errors";
 import { mkAlgobEnv } from "../helpers/params";
 import { cleanupMutableData } from "../lib/script-checkpoints";
