+++ packages/algob/test/mocks/account.ts
@@ -1,6 +1,5 @@
+import type { Account } from "@algorand-builder/runtime/build/types";
 import { mnemonicToSecretKey } from "algosdk";
-
-import { Account } from "../../src/types";
 
 export const mnemonic1 = "call boy rubber fashion arch day capable one sweet skate outside purse six early learn tuition eagle love breeze pizza loud today popular able divide";
 const a1 = mnemonicToSecretKey(mnemonic1);
