+++ packages/algob/test/mocks/deployer.ts
@@ -1,9 +1,8 @@
-import type { SSCDeploymentFlags, SSCOptionalFlags, TxParams } from "@algorand-builder/runtime/build/types";
+import type { Account, SSCDeploymentFlags, SSCOptionalFlags, TxParams } from "@algorand-builder/runtime/build/types";
 import type { LogicSig, LogicSigArgs } from "algosdk";
 import * as algosdk from "algosdk";
 
 import type {
-  Account,
   AlgobDeployer,
   ASADeploymentFlags,
   ASAInfo,
