+++ packages/runtime/src/index.ts
@@ -1,6 +1,9 @@
 import { StoreAccount } from "./account";
-import { applyErrorMessageTemplate } from "./errors/errors";
+import { applyErrorMessageTemplate, BuilderError } from "./errors/errors";
+import { parseZodError } from "./errors/validation-errors";
 import { Interpreter } from "./interpreter/interpreter";
+import { loadASAFile, validateASADefs } from "./lib/asa";
+import { loadFromYamlFileSilent, loadFromYamlFileSilentWithMessage } from "./lib/files";
 import { addressToPk, parseSSCAppArgs, stringToBytes, uint64ToBigEndian } from "./lib/parsing";
 import { encodeNote, mkTransaction } from "./lib/txn";
 import { Runtime } from "./runtime";
@@ -9,11 +12,17 @@ export {
   Interpreter,
   Runtime,
   StoreAccount,
+  BuilderError,
   mkTransaction,
   applyErrorMessageTemplate,
   parseSSCAppArgs,
   addressToPk,
   uint64ToBigEndian,
   stringToBytes,
-  encodeNote
+  encodeNote,
+  loadFromYamlFileSilent,
+  loadFromYamlFileSilentWithMessage,
+  loadASAFile,
+  parseZodError,
+  validateASADefs
 };
