+++ packages/runtime/src/types.ts
@@ -1,11 +1,12 @@
 import {
-  Account,
+  Account as AccountSDK,
   AssetDef,
   AssetHolding,
   LogicSig,
   SSCSchemaConfig,
   TxnEncodedObj
 } from "algosdk";
+import * as z from 'zod';
 
 import {
   Add, Addr, Arg, Byte, Bytec, Bytecblock, Div, Int, Len, Mul, Pragma,
@@ -13,6 +14,7 @@ import {
 } from "./interpreter/opcode-list";
 import { TxnFields } from "./lib/constants";
 import type { IStack } from "./lib/stack";
+import type { ASADefSchema, ASADefsSchema } from "./types-input";
 
 export type Operator = Len | Add | Sub |
 Mul | Div | Arg | Bytecblock | Bytec | Addr | Int | Byte | Pragma;
@@ -117,7 +119,7 @@ export interface StoreAccountI {
   appsTotalSchema: SSCSchemaConfig
   createdApps: Map<number, SSCAttributesM>
   createdAssets: Map<number, AssetDef>
-  account: Account
+  account: AccountSDK
 
   balance: () => number
   getApp: (appId: number) => SSCAttributesM | undefined
@@ -168,7 +170,7 @@ export interface TxParams {
 /**
  * Stateful Smart contract flags for specifying sender and schema */
 export interface SSCDeploymentFlags extends SSCOptionalFlags {
-  sender: Account
+  sender: AccountSDK
   localInts: number
   localBytes: number
   globalInts: number
@@ -210,7 +212,7 @@ export interface Sign {
 
 export interface AlgoTransferParam extends Sign {
   type: TransactionType.TransferAlgo
-  fromAccount: Account
+  fromAccount: AccountSDK
   toAccountAddr: AccountAddress
   amountMicroAlgos: number
   payFlags: TxParams
@@ -218,7 +220,7 @@ export interface AlgoTransferParam extends Sign {
 
 export interface AssetTransferParam extends Sign {
   type: TransactionType.TransferAsset
-  fromAccount: Account
+  fromAccount: AccountSDK
   toAccountAddr: AccountAddress
   amount: number
   assetID: number
@@ -228,7 +230,7 @@ export interface AssetTransferParam extends Sign {
 export interface SSCCallsParam extends SSCOptionalFlags, Sign {
   type: TransactionType.CallNoOpSSC | TransactionType.ClearSSC |
   TransactionType.CloseSSC | TransactionType.DeleteSSC
-  fromAccount: Account
+  fromAccount: AccountSDK
   appId: number
   payFlags: TxParams
 }
@@ -236,3 +238,15 @@ export interface SSCCallsParam extends SSCOptionalFlags, Sign {
 export interface AnyMap {
   [key: string]: any // eslint-disable-line @typescript-eslint/no-explicit-any
 }
+
+export interface Account extends AccountSDK {
+  // from AccountSDK: addr: string;
+  //                  sk: Uint8Array
+  name: string
+}
+
+export type AccountMap = Map<string, Account>;
+
+export type ASADef = z.infer<typeof ASADefSchema>;
+
+export type ASADefs = z.infer<typeof ASADefsSchema>;
